backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
httpx>=0.24.0
redis>=4.5.5
fastapi-cache2>=0.2.0
python-dotenv>=0.21.0
pydantic>=1.10.4